pyproject.toml
[tool.poetry]
name = "funnel"
version = "0.1.0"
description = ""
authors = ["Vasiliynemix <vasyanemix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.2.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
black = "^23.11.0"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
